Consolidate wallet component package
