+++ packages/ui-components/src/components/Wallet/AddressInput.test.tsx
@@ -1,11 +1,11 @@
 import {fireEvent, waitFor} from '@testing-library/react';
 import React from 'react';
 
-import * as domainProfileActions from '../../../../actions/domainProfileActions';
-import useResolverKeys from '../../../../hooks/useResolverKeys';
-import {VALID_ETH_ADDRESS} from '../../../../tests/common';
-import {mockProfileData} from '../../../../tests/mocks/wallet';
-import {customRender} from '../../../../tests/test-utils';
+import * as domainProfileActions from '../../actions/domainProfileActions';
+import useResolverKeys from '../../hooks/useResolverKeys';
+import {VALID_ETH_ADDRESS} from '../../tests/common';
+import {mockProfileData} from '../../tests/mocks/wallet';
+import {customRender} from '../../tests/test-utils';
 import AddressInput from './AddressInput';
 
 const defaultProps = {
