+++ packages/ui-components/src/components/Wallet/Buy.tsx
@@ -6,10 +6,10 @@ import React from 'react';
 import config from '@unstoppabledomains/config';
 import {makeStyles} from '@unstoppabledomains/ui-kit/styles';
 
-import type {SerializedWalletBalance} from '../../../../lib';
-import {useTranslationContext} from '../../../../lib';
-import type {TokenEntry} from '../../../Wallet/Token';
+import type {SerializedWalletBalance} from '../../lib';
+import {useTranslationContext} from '../../lib';
 import {SelectAsset} from './SelectAsset';
+import type {TokenEntry} from './Token';
 
 const useStyles = makeStyles()((theme: Theme) => ({
   flexColCenterAligned: {
