+++ packages/ui-components/src/components/Wallet/InlineEducation.tsx
