--- packages/ui-components/src/components/Wallet/index.ts
@@ -1,4 +1,0 @@
-export * from './DomainWalletList';
-export * from './DomainWalletTransactions';
-export * from './TokensPortfolio';
-export {default as AccessWallet} from './AccessWallet';
