+++ packages/ui-components/src/components/Wallet/index.tsx
@@ -4,8 +4,8 @@ import React from 'react';
 
 import {makeStyles} from '@unstoppabledomains/ui-kit/styles';
 
-import {isDomainValidForManagement} from '../../../../lib';
-import type {ManageTabProps} from '../../common/types';
+import {isDomainValidForManagement} from '../../lib';
+import type {ManageTabProps} from '../Manage/common/types';
 import {Configuration} from './Configuration';
 import {Header} from './Header';
 
@@ -15,6 +15,11 @@ const useStyles = makeStyles()((theme: Theme) => ({
     flexDirection: 'column',
   },
 }));
+
+export * from './DomainWalletList';
+export * from './DomainWalletTransactions';
+export * from './TokensPortfolio';
+export {default as AccessWallet} from './AccessWallet';
 
 export const Wallet: React.FC<
   ManageTabProps & {
